requirements.txt
setuptools
numpy
pandas
google-cloud-storage
pyyaml
tensorflow
python-dotenv
joblib
scikit-learn
comet-ml
dvc
dvc-gs
flask